Fix window focus, bind events to ownerDocument
We used target.ownerDocument.documentElement for event binding, since
servo doesn't yet bubble events to window. However,
target.ownerDocument.documentElement does not receive window focus/blur
events, which we need for windowcontrols.

This commit switches to target.ownerDocument for event binding in the
hopes that this will work for Servo. I'm not sure how to test that.

cc @paulrouget