Add type annotations for driver module.
With this change flow will be able to ensure type check uses of `on(address, eventToAction)` or more specifically it will ensure that `eventToAction` returns `action` that is compatible with a component receiving it.

This uncovered few existing bugs that this patch also fixes.

Special `driver` module had to get renamed to `@driver`, because `.flowconfig` mapping was incorrectly mapping stuff in node_modules that contained `driver`.