Create module for managing animations
This is successor to a Stopwatch module that unlike it predecessor as it keeps track of current state & there for is able to switch animations from the specific state.

For example if if opacity was animating from 0 to 1 and animation switched while opacity was 0.4 to animate back to 0 it will start new transition from 0.4 unlike 1 which was case for Stopwatch.