Adapt Overlay module
Overlay used to serve two different purposes in the past. One to overlay the WebView when sidebar was open so that clicking it would bring back web-view & the other to display shadow under the Suggestions.

After things moved around it made more sense to have two different overlays. One that can cover Navigators when sidebar is expanded and other that provides shadow under the Suggestions. Former is single item while the later is per tab.

Fading animation is now rewritten with new Animation module.