Workaround iframe API limitations at the DOM level. (#1228)
We used to keep track `src` and currently loaded `url` as separate fields in the model so that we would only update `src` when user edited url. But that logic had a major issue #949:

If user initial loads url `a` than navigates to `b` by clicking a link and then edits url to navigate to `a` nothing would happen. That is because Virtual DOM would diff and see that src in both cases is `a` and do not update `iframe.src`.

With this change we use `setting(location, url)` for updating `iframe.src`. As of `location` setter it keeps track of iframe location and only reflects passed `url` if it is different from it.

This approach fixes #949 and makes state model simpler.